Remove unwanted POM manipulation functionality
By default, this extension manipulates other parts of the pom file
other than the version. This includes adding additional plugin
configurations and adding a comment at the bottom of the pom file
indicating it was automatically updated. The plugin manipulation can
be explicitly disabled at runtime with -D parameters. However, since
we don't ever want this, it's cleaner just to remove this
functionality.